pyproject.toml
[project]
name = "information_retrieval_rag"
version = "0.1.0"
description = "Modular RAG infra for experiments and submissions"
authors = [
    {name = "Lee Sang-Won", email = "refelicity6077@gmail.com"},
    {name = "Kang Yeon-Kyung", email = "yeon97@gmail.com"},
    {name = "Kim Hyo-Seok", email = "hskimh1982@gmail.com"},
    {name = "Jung Jae-Hoon", email = "rpajhc@gmail.com"},
    {name = "Choi Woong-Bi", email = "wchoi189@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "uvicorn[standard]>=0.21.1,<0.22.0",
    "elasticsearch>=8.8.0,<9.0",
    "transformers>=4.56.1,<5.0",
    "requests>=2.31.0,<3.0",
    "sentencepiece>=0.2.1,<0.3",
    "scikit-learn>=1.7.1,<2.0",
    "openai>=1.106.1,<2.0",
    "torch>=2.6.0,<3.0",
    "numpy>=1.26.0,<2.0",
    "omegaconf>=2.0,<3.0",
    "python-dotenv>=1.0.0,<2.0",
    "hydra-core>=1.3,<2.0",
    "wandb>=0.21.3,<0.22",
    "redis>=4.5.5,<5.0",
    "icecream>=2.1.2,<3.0",
    "fire>=0.5.0,<0.6",
    "pytest>=7.4,<8.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "Jinja2>=3.1.4,<4.0",
    "rich>=13.7.0,<14.0",
    "questionary>=2.0.1,<3.0",
    "matplotlib>=3.10.6,<4.0",
    "seaborn>=0.13.2,<0.14",
    "plotly>=6.3.0,<7.0",
    "kaleido>=1.1.0,<2.0",
    "instructor>=1.11.3,<2.0",
    "fastapi>=0.116.1,<0.117",
    "streamlit>=1.49.1,<2.0",
    "pandas>=2.3.2,<3.0",
    "accelerate>=1.10.1,<2.0",
    "bitsandbytes>=0.41.0,<0.42",
    "nltk>=3.8.1,<4.0",
    "tqdm>=4.67.1,<5.0",
    "sacremoses>=0.1.1,<0.2",
    "datasets>=4.0.0,<5.0",
    "optimum>=1.20.0,<2.0",
    "googletrans==4.0.2",
    "ruamel-yaml>=0.18.15,<0.19",
    "sentence-transformers>=5.1.0,<6.0",
    "typer>=0.17.4,<0.18",
    "kiwipiepy>=0.21.0,<0.22"
]
[project.scripts]
evaluate = "scripts.evaluate:main"
cli = "scripts.cli_menu:main"

[project.optional-dependencies]
dev = [
    "black>=24.3,<25.0",
    "ruff>=0.7,<0.8",
    "tqdm>=4.66.1,<5.0",
    "pytest-asyncio==0.21.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--ignore=scripts/integration/ --ignore=scripts/debugging/"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
